Extract habit streak lists & use adaptive header; refactoring
