+++ lib/l10n/app_en.arb
@@ -73,8 +73,10 @@
   "habitsCompletedHeader": "Completed",
   "habitsOpenHeader": "Open",
   "habitsPendingLaterHeader": "Pending later",
-  "habitsShortStreaksHeader": "3-day streaks (or longer)",
   "habitsLongerStreaksHeader": "Week-long streaks (or longer)",
+  "habitsLongerStreaksEmptyHeader": "No week-long streaks at the moment",
+  "habitsShortStreaksHeader": "Three day streaks (or longer)",
+  "habitsShortStreaksEmptyHeader": "No three day streaks at the moment",
   "journalDateFromLabel": "Date from:",
   "journalDateInvalid": "Invalid Date Range",
   "journalDateNowButton": "now",
